Addjusted size parameters in API functions to be of type size_t
This change allows for users to use sizeof and not worry about overflows.
The size_t type is guaranteed to be able to access all array positions in the language definition.

+++ include/fti-intern.h
@@ -363,7 +363,7 @@ extern "C" {
      */
     typedef struct FTIT_type {
         int id;                              /**< ID of the data type.        */
-        int size;                            /**< Size of the data type.      */
+        size_t size;                         /**< Size of the data type.      */
         FTIT_complexType* structure;         /**< Logical structure for HDF5. */
         FTIT_H5Group* h5group;               /**< Group of this datatype.     */
 #ifdef ENABLE_HDF5
